backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "investpulse-backend"
version = "0.1.0"
description = "InvestPulse backend processing system"
requires-python = ">=3.8"
dependencies = [
    "watchdog",
    "requests",
    "pandas",
    "pyarrow",
    "sec-edgar-downloader",
    "beautifulsoup4",
    "pyspark",
    "ray[default]",
    "mypy",
    "pandas-stubs",
    "types-requests",
    "types-html5lib",
    "lxml-stubs",
    "pytest",
    "pytest-asyncio",
    "pytest-dotenv",
    "python-dotenv",
    "tweepy>=4.14.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "sentencepiece>=0.1.99",
    "aiohttp>=3.9.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
env_files = [
    "../secrets/.env"
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
show_error_codes = true
show_column_numbers = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_any_generics = true
disallow_subclassing_any = true
check_untyped_defs = true
strict_optional = true
no_implicit_optional = true
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = ["watchdog.*", "sec_edgar_downloader.*", "bs4.*", "pyarrow.*", "pyspark.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
